Move Flag and Axis into Components
