+++ src/Charts/ScatterPlot.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/no-unknown-property */
 import { h, Component, VNode } from 'preact';
-import { Axis } from '../Axis';
+import { Axis } from '../Components/Axis';
 import { Margin, DataArray, NumberObject } from '../types';
 import { scaleLinear } from 'd3-scale';
 import { extent } from 'd3-array';
